# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ModelInfo {
  id: string;
  label: string;
  provider: string;
}

export interface ModelsResponse {
  models: ModelInfo[];
  defaultModel: string;
}

export interface SeedResponse {
  meetingTitle: string;
  date: string;
  participants: string[];
  v1Transcript: string;
  v1Summary: string;
  feedback: string;
  models: ModelInfo[];
  defaultModel: string;
}

export interface RegenerateRequest {
  /** @minLength 1 */
  v1Transcript: string;
  /** @minLength 1 */
  v1Summary: string;
  /** @minLength 1 */
  feedback: string;
  flaggedSegments?: string[];
  /** @minLength 1 */
  model: string;
}

export type DiffSegmentType = typeof DiffSegmentType[keyof typeof DiffSegmentType];


export const DiffSegmentType = {
  same: 'same',
  add: 'add',
  remove: 'remove',
} as const;

export interface DiffSegment {
  type: DiffSegmentType;
  value: string;
}

export interface DiffStats {
  added: number;
  removed: number;
  unchanged: number;
}

export interface Observability {
  model: string;
  inputTokens: number;
  outputTokens: number;
  totalTokens: number;
  latencyMs: number;
  estimatedCostUsd: number | null;
  finishReason: string;
  reasoning: string | null;
}

export interface RegenerateResult {
  v2Transcript: string;
  v2Summary: string;
  changeExplanation: string;
  transcriptDiff: DiffSegment[];
  summaryDiff: DiffSegment[];
  transcriptDiffStats: DiffStats;
  summaryDiffStats: DiffStats;
  observability: Observability;
}

export interface ErrorResponse {
  error: string;
  retryable: boolean;
}