# lib/api-zod/src/generated/types/diffSegment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DiffSegmentType } from './diffSegmentType';

export interface DiffSegment {
  type: DiffSegmentType;
  value: string;
}