# lib/api-zod/src/generated/types/regenerateRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegenerateRequest {
  /** @minLength 1 */
  v1Transcript: string;
  /** @minLength 1 */
  v1Summary: string;
  /** @minLength 1 */
  feedback: string;
  flaggedSegments?: string[];
  /** @minLength 1 */
  model: string;
}