# lib/api-zod/src/generated/types/regenerateResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DiffSegment } from './diffSegment';
import type { DiffStats } from './diffStats';
import type { Observability } from './observability';

export interface RegenerateResult {
  v2Transcript: string;
  v2Summary: string;
  changeExplanation: string;
  transcriptDiff: DiffSegment[];
  summaryDiff: DiffSegment[];
  transcriptDiffStats: DiffStats;
  summaryDiffStats: DiffStats;
  observability: Observability;
}